Add Stage 4/6 project
Now the program allows to buy one cup of coffee (several types) or to fill the supplies or to take its money out. The program does one of the mentioned actions at a time. It also calculates how many ingredients and money have left, and displays the number of supplies before and after an action.